requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==3.2
django-debug-toolbar==3.4.0
django-split-settings==1.2.0
Faker==16.6.1
flake8==4.0.0
mccabe==0.6.1
psycopg2-binary==2.9
pycodestyle==2.10.0
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
